reset selected file after deleting

+++ client/modules/IDE/actions/files.js
@@ -157,23 +157,32 @@ export function updateFileName(id, name) {
 export function deleteFile(id, parentId) {
   return (dispatch, getState) => {
     const state = getState();
-    const deleteConfig = {
-      params: {
-        parentId
-      }
-    };
-    axios.delete(`${ROOT_URL}/projects/${state.project.id}/files/${id}`, deleteConfig, { withCredentials: true })
-      .then(() => {
-        dispatch({
-          type: ActionTypes.DELETE_FILE,
-          id,
+    if (state.project.id) {
+      const deleteConfig = {
+        params: {
           parentId
+        }
+      };
+      axios.delete(`${ROOT_URL}/projects/${state.project.id}/files/${id}`, deleteConfig, { withCredentials: true })
+        .then(() => {
+          dispatch({
+            type: ActionTypes.DELETE_FILE,
+            id,
+            parentId
+          });
+        })
+        .catch(response => {
+          dispatch({
+            type: ActionTypes.ERROR,
+            error: response.data
+          });
         });
-      })
-      .catch(response => dispatch({
-        type: ActionTypes.ERROR,
-        error: response.data
-      })
-    );
+    } else {
+      dispatch({
+        type: ActionTypes.DELETE_FILE,
+        id,
+        parentId
+      });
+    }
   };
 }
